*** empty log message ***

+++ AlephOneSrc/RenderPlaceObjs.cpp
@@ -128,7 +128,7 @@ void RenderPlaceObjsClass::build_render_object_list()
 			sorted_node_data *base_nodes[MAXIMUM_OBJECT_BASE_NODES];
 			
 			//float Opacity = 1;(object_index == self_index) ? GetChaseCamData().Opacity : 1;
-			float Opacity = (object_index == self_index) ? 0.4 : 1;
+			float Opacity = (float)((object_index == self_index) ? 0.4 : 1);
 			render_object_data *render_object= build_render_object(NULL, floor_intensity, ceiling_intensity,
 				base_nodes, &base_node_count, object_index, Opacity);
